Removed the microphone example from astrobee api.
The microphone example was an APK for the ISAAC project. It was moved
out of the astrobee api project and into its own project. The astrobee
api failed to compile because the microphone example was still in the
gradle settings file. It has been removed and the astrobee api compiles
now.