Fix astrobee api scripts (#4)
* Changed the command file generation scripts.

Astrobee common has been integrated into the astrobee repo. The command
file generation scripts for the astrobee api use some of the files that
were integrated. These scripts have been updated to reflect the correct
file location.

* Removed the microphone example from astrobee api.

The microphone example was an APK for the ISAAC project. It was moved
out of the astrobee api project and into its own project. The astrobee
api failed to compile because the microphone example was still in the
gradle settings file. It has been removed and the astrobee api compiles
now.

* Changed default location of the Android repo.

The astrobee api scripts now assume that the android repo is located in
the astrobee fsw submodules folder. These scripts are mostly used by
internal developers so it makes more sense that the default location
matches the internal repo setup.

Co-authored-by: Katie Browne <kathryn.browne@nasa.gov>