Modified test gs apk to not un/dock.
The test simple trajectory gs apk commands the robot to undock, do a
simple trajectory, and then dock if the robot is operating in the
granite or iss world. The iss simulated world starts with the robot
docked but the granite world does not so the apk has been changed to
only do the simple trajectory in the granite world.