Fixed science camera resize image bug.
The bug that caused the images not to be resized before being published
to ros has been fixed. The default size of the published image has been
set to the full size since recording the full size compressed image in
a bag didn't cause any run time issues. With publishing the full size
images by default, the facility will only need to download the bags for
their panorama activities.