Update gs_developer_guide.md
This is a minor change, since the latest version of Android Studio, from what I have been told, does not work well with astrobee_android. And I know that it works on this version, so to save time on figuring out which versions work or don't work, it's easier to use this version that does work.